stripe nccl envs from benchmark
